Add Horse API. Adds BUKKIT-4424
API has been added to interface with Horses and to modify their inventories.

Previously, all Tameable animals have been animals capable of sitting. However, with the introduction of Horses in 1.6, this is no longer the case. As a result, sitting functions have been moved to their own interface so that Ocelot and Wolfs may specifically extend them. 

This commit fixes BUKKIT-4393.